+++ src/controllers/posts.py
@@ -1,6 +1,6 @@
 import uuid
 
-from src.database import SessionLocal
+from src.database.connection import SessionLocal
 from src.database.models.posts import Posts
 from src.schemas.posts import PostsModel
